Emily's first pass
Hi- I did a first pass through here. I think after reading through this. There are numerous comments, set off with <!--EM. Please review/respond. 

One thing that I think we need to do is to take out Step 1 in this piece and create a separate piece that all of our RHM content points to about Configuring an OpenShift Cluster on RHM. The content is identical to the first steps here: https://developer.ibm.com/tutorials/get-started-using-a-cockroachdb-operator-hosted-on-red-hat-marketplace/ I think it might be better to just make it separate piece so that we can point to it in every piece of contetn without hurting our SEO (which would happen if we have the same content in the top section of each piece.