Emily's second pass
I took out the duplicate content and pointed to the tutorial in a prereqs section. Do you think that's enough or do you think I need to add it as Step 1. "Configure a Red Hat OpenShift cluster on Red Hat Marketplace" and the say: The instructions to do this are in this tutorial: [link to the tutorial].

A few other questions:
1. This line says, "check the logs and be sure there are no errors in the Ansible script" . . . does the reader need to have Ansible installed or does that happen automatically with an OpenShift install?
2. Under Step 3. Access the cluster on your localhost, you mentioned pgAdmin 4 console . . . . what is pgAdmin 4 console? Is that part of the marketplace? how does a reader access that?